fix(cli): separate handler vite handler
